requirements.txt
Flask==2.0.1
solana==0.0.4
websockets==11.0.3
python-dotenv==0.19.0
cryptography==36.0.1
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
python-jose==3.3.0
pyjwt==2.1.0
python-telegram-bot==20.7
python-x-sign==1.0.0
aiohttp==3.9.1
base58==2.1.1
PyNaCl==1.5.0
python-socketio==4.6.1
flask-socketio==4.3.1
